:heavy_plus_sign: Add implementation of fetch tree data in tree component
